fix: wallet-related stuff still on support config screen
